requirements.txt
fastapi[standard]
Pillow
python-multipart
torch
transformers
uvicorn